AppWeb-Here/AppWeb-Here/web-asistencia-main/face-recognition-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy
opencv-python
face-recognition
Pillow
python-multipart
pydantic==2.5.0
requests==2.31.0
